# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DiaryIn": {
        "additionalProperties": false,
        "properties": {
          "is_sensitive": {
            "default": false,
            "title": "Is Sensitive",
            "type": "boolean"
          },
          "tags": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "text": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "time": {
            "$ref": "#/components/schemas/EventTimeIn"
          }
        },
        "required": [
          "text",
          "time"
        ],
        "title": "DiaryIn",
        "type": "object"
      },
      "DiaryOut": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "const": "fact",
            "default": "fact",
            "title": "Category",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_sensitive": {
            "title": "Is Sensitive",
            "type": "boolean"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProvenanceOut"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "time": {
            "$ref": "#/components/schemas/EventTimeOut"
          }
        },
        "required": [
          "id",
          "text",
          "is_sensitive",
          "tags",
          "time",
          "provenance"
        ],
        "title": "DiaryOut",
        "type": "object"
      },
      "DiaryPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DiaryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMetadata"
          },
          "revisions": {
            "items": {
              "$ref": "#/components/schemas/DiaryOut"
            },
            "title": "Revisions",
            "type": "array"
          }
        },
        "required": [
          "items",
          "revisions",
          "page"
        ],
        "title": "DiaryPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "HealthCurve",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/diary-events": {
      "get": {
        "operationId": "list_diary_api_v1_diary_events_get",
        "parameters": [
          {
            "description": "Sensitive entries are excluded from default views (T7).",
            "in": "query",
            "name": "include_sensitive",
            "required": false,
            "schema": {
              "default": false,
              "description": "Sensitive entries are excluded from default views (T7).",
              "title": "Include Sensitive",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Diary",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "create_diary_api_v1_diary_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Diary",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/life-events": {
      "get": {
        "operationId": "list_life_events_api_v1_life_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_sensitive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Sensitive",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifeEventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Life Events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "create_life_event_api_v1_life_events_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LifeEventIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifeEventOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Life Event",
        "tags": [
          "events"
        ]
      }
    },
...
    "/api/v1/symptoms": {
      "get": {
        "operationId": "list_symptoms_api_v1_symptoms_get",
        "parameters": [
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SymptomPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Symptoms",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "create_symptom_api_v1_symptoms_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SymptomIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SymptomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Symptom",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/symptoms/{event_id}/correct": {
      "post": {
        "operationId": "correct_symptom_api_v1_symptoms__event_id__correct_post",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "hc_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hc Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SymptomCorrectionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SymptomOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correct Symptom",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}